Use floating point precision for tax rates

+++ BaseLinkerApi/Requests/ExternalStorages/GetExternalStorageProductsData.cs
@@ -63,7 +63,7 @@ namespace BaseLinkerApi.Requests.ExternalStorages
             public decimal PriceWholesaleNetto { get; set; }
 
             [JsonPropertyName("tax_rate")]
-            public int TaxRate { get; set; }
+            public double TaxRate { get; set; }
             
             [JsonPropertyName("weight")]
             public double Weight { get; set; }
